Added a program for making the movie.
To use it, open the Serial Monitor and type in, "red", "green", "blue",
"off", or "clear" (this is only if the program isn't responding).
Typing in one of the colors will fade the chair over 3 seconds to the
new color from the color it was previously or from off. Typing off
fades the LEDs out over 5 seconds.